ml/requirements.txt
fastapi
redis
types-redis
uvicorn
scikit-learn
pandas
pillow
numpy
joblib
